# prueba-bex-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Visitas con Autenticación",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "Iniciar sesión",
                "description": "Intenta autenticar al usuario con el nombre de usuario y la contraseña proporcionados. Si la autenticación es exitosa, devuelve un token de autenticación.",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Nombre de usuario del usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticación exitosa, retorna el token JWT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Token JWT de autenticación",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas, no autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Credenciales incorrectas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Detalles del error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/visits": {
            "get": {
                "summary": "Obtener lista de visitas",
                "description": "Obtener una lista de visitas paginada",
                "operationId": "59bc502e31c31bcb9d32d86e06260de5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de visitas paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "records": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Visit"
                                            }
                                        },
                                        "totalPages": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "totalRecords": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "summary": "Crear una nueva visita",
                "description": "Crear una nueva visita",
                "operationId": "2ff60d5091a706b6942379521ae9074f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Visit"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Visita creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/visits/{id}": {
            "get": {
                "summary": "Mostrar una visita específica",
                "description": "Obtener los detalles de una visita específica por su ID",
                "operationId": "42d6ce6dce87180d7400f1a2ad20259e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la visita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la visita",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visita no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Actualizar una visita",
                "description": "Actualizar los detalles de una visita específica",
                "operationId": "460d80e475b87e62bf034f944d29cacf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la visita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Visit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visita actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visit"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Visita no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Eliminar una visita",
                "description": "Eliminar una visita específica por su ID",
                "operationId": "17961e50292ffea006c5ff414d95307b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la visita",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visita eliminada correctamente"
                    },
                    "404": {
                        "description": "Visita no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Visit": {
                "title": "Visit",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Visit 1"
                    },
                    "email": {
                        "type": "string",
                        "example": "email@email.com"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Por favor ingresa tu token JWT en el campo de autorización.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}